Wip functional component for creating a post.
