+++ public/glue.js
@@ -1,9 +1,5 @@
 const invoke = window.__TAURI__.invoke
 
-export async function invokeGreet(name) {
-    return await invoke("greet", {name: name})
-}
-
 export async function invokeCreatePost(user_id, body) {
     return await invoke("create_post", {user_id: user_id, body: body});
 }
