+++ src/app.rs
@@ -1,4 +1,5 @@
-use crate::components::footer::{Footer, Props};
+use crate::components::create_post::CreatePost;
+use crate::components::footer::Footer;
 use serde::{Deserialize, Serialize};
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::spawn_local;
@@ -6,19 +7,8 @@ use yew::prelude::*;
 
 #[wasm_bindgen(module = "/public/glue.js")]
 extern "C" {
-    #[wasm_bindgen(js_name = invokeGreet)]
-    async fn greet(name: String) -> JsValue;
-
-    #[wasm_bindgen(js_name = invokeCreatePost, catch)]
-    async fn create_post(user_id: String, body: String) -> Result<JsValue, JsValue>;
-
     #[wasm_bindgen(js_name = invokeGetPost, catch)]
     async fn get_post(id: String) -> Result<JsValue, JsValue>;
-}
-
-#[derive(Serialize, Deserialize)]
-struct GreetArgs<'a> {
-    name: &'a str,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -28,97 +18,8 @@ struct GetPostArgs<'a> {
 
 #[function_component(App)]
 pub fn app() -> Html {
-    let greet_input_ref = use_node_ref();
+    // GET POST
 
-    let name = use_state(|| String::new());
-
-    let greet_msg = use_state(|| String::new());
-    {
-        let greet_msg = greet_msg.clone();
-        let name = name.clone();
-        let name2 = name.clone();
-        use_effect_with_deps(
-            move |_| {
-                spawn_local(async move {
-                    if name.is_empty() {
-                        return;
-                    }
-                    // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
-                    let new_msg: String = greet(name.to_string()).await.as_string().unwrap();
-                    greet_msg.set(new_msg);
-                });
-
-                || {}
-            },
-            name2,
-        );
-    }
-
-    let greet = {
-        let name = name.clone();
-        let greet_input_ref = greet_input_ref.clone();
-        Callback::from(move |e: SubmitEvent| {
-            e.prevent_default();
-            name.set(
-                greet_input_ref
-                    .cast::<web_sys::HtmlInputElement>()
-                    .unwrap()
-                    .value(),
-            );
-        })
-    };
-
-    /// CREATE POST
-    let create_post_body_input_ref = use_node_ref();
-    let create_post_body = use_state(|| String::new());
-
-    let create_post_msg = use_state(|| String::new());
-    {
-        let create_post_msg = create_post_msg.clone();
-        let create_post_body = create_post_body.clone();
-        let create_post_body2 = create_post_body.clone();
-        use_effect_with_deps(
-            move |_| {
-                spawn_local(async move {
-                    if create_post_body.is_empty() {
-                        return;
-                    }
-                    // let args = to_value(&GetPostArgs { id: &*post_id }).unwrap();
-                    // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
-                    let result: Result<JsValue, JsValue> = create_post(
-                        "8b6a5f15-7e5c-4418-b3c3-672f5c24abb7".to_string(),
-                        create_post_body.to_string(),
-                    )
-                    .await;
-                    let new_msg = match result {
-                        Ok(response) => response.as_string().unwrap(),
-                        Err(error) => error.as_string().unwrap(),
-                    };
-                    create_post_msg.set(new_msg);
-                });
-
-                || {}
-            },
-            create_post_body2,
-        );
-    }
-
-    let create_post = {
-        println!("submitted request to create post");
-        let create_post_body = create_post_body.clone();
-        let create_post_body_input_ref = create_post_body_input_ref.clone();
-        Callback::from(move |e: SubmitEvent| {
-            e.prevent_default();
-            create_post_body.set(
-                create_post_body_input_ref
-                    .cast::<web_sys::HtmlInputElement>()
-                    .unwrap()
-                    .value(),
-            );
-        })
-    };
-
-    /// CREATE POST
     let get_post_input_ref = use_node_ref();
 
     let post_id: UseStateHandle<String> = use_state(|| String::new());
@@ -167,13 +68,6 @@ pub fn app() -> Html {
     html! {
         <main class="container">
 
-            <form class="row" onsubmit={greet}>
-                <input id="greet-input" ref={greet_input_ref} placeholder="Enter a name..." />
-                <button type="submit">{"Greet"}</button>
-            </form>
-
-            <p><b>{ &*greet_msg }</b></p>
-
             <form class="row" onsubmit={get_post}>
                 <input id="get-post-input" ref={get_post_input_ref} placeholder="Enter the post id to fetch" />
                 <button type="submit">{"Submit"}</button>
@@ -181,12 +75,7 @@ pub fn app() -> Html {
 
             <p><b>{ &*get_post_msg }</b></p>
 
-            <form class="row" onsubmit={create_post}>
-                <input id="get-post-input" ref={create_post_body_input_ref} placeholder="Enter the post message" />
-                <button type="submit">{"Submit"}</button>
-            </form>
-
-            <p><b>{ &*create_post_body }</b></p>
+            <CreatePost />
 
             <Footer />
         </main>
