+++ src/components/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod create_post;
 pub(crate) mod footer;
